Move musicbrainz to beetsplug directory
